now run in a subprocess
